requirements.txt
anthropic
beautifulsoup4
feedparser
openai
pydub
python-dotenv
requests
tavily-python
youtube-transcript-api
